refactor(Notifications): ♻️ Refactored Entity mappings, Service and Controller
now The controller returns NotificationResponse
The service now validates the FK userId to exist on notification creation

+++ src/main/java/com/baio/money_minder/repositories/NotificationRepository.java
@@ -8,4 +8,4 @@ import org.springframework.stereotype.Component;
 import java.util.Optional;
 
 @Component
-public interface NotificationRepository extends JpaRepository<Notification, Integer> { }
+public interface NotificationRepository extends JpaRepository<Notification, Long> { }
